requirements.txt
ansicon==1.89.0
arrow==1.0.3
asgiref==3.3.1
Babel==2.9.0
bert-tensorflow==1.0.4
blessed==1.18.0
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.8
dateutils==0.6.12
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-heroku==0.3.1
django-jsonresponse==0.10.0
django-mathfilters==1.0.0
django-otp==1.0.3
django-phone-field==1.8.1
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-picklefield==3.0.1
django-q==1.3.5
django-reset-migrations==0.4.0
django-rest-framework==0.1.0
django-rest-framework-braces==0.3.4
django-sendsms==0.2.3
django-sharing==0.0.2
django-snippetscream==0.0.7
djangorestframework==3.12.4
flutter==0.1
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
importlib-metadata==3.10.0
jinxed==1.0.1
Markdown==3.3.4
phonenumberslite==8.12.21
Pillow==8.1.2
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.56.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1
